% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timely.R
\name{timely}
\alias{timely}
\title{Timely: Generate a time-stamped filename}
\usage{
timely(name = "test", extension = "csv", time = "\%m\%d\%y_\%H.\%M")
}
\arguments{
\item{name}{a string with the title of the file}

\item{extension}{a file extension, e.g. csv or dta}

\item{time}{optional timestamp defintion, see format() for syntax}
}
\value{
a string
}
\description{
Timely: Generate a time-stamped filename
}
\examples{
timely()
timely("data_output", "dta", "\%m\%d\%y_\%H.\%M")
}
